backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
supabase
requests
joblib
pydantic
python-dotenv